// src/components/air-pricing/interface.ts
export interface AirPricingDetailDTOs {
  [key: string]: string;
}
export interface AirPricingFeeDTOs {
  airPricingFeeID: string;
  feeID: string;
  feeName: string;
  feeNo: string;
  currencyID: string;
  currencyName: string;
  unitID: string;
  internationalCode: string;
  price: string;
}
// ...
export interface FormValues {
  airPricingID: string;
  aodid: string;
  aolid: string;
  commodityID: string;
  note: string;
  dateEffect: Dayjs;
  validityDate: Dayjs;
  freqDate: string;
  demAirPricing: string;
  detAirPricing: string;
  stoAirPricing: string;
  lclMinAirPricing: string;
  currencyID: string;
  lclAirPricing: string;
  public: boolean;
  statusAirPricing: string;
  airPricingDetailDTOs: AirPricingDetailDTOsFormValue[];
  airPricingFeeDTOs: AirPricingFeeFormValue[];
}

export interface AirPricingDetailDTOsFormValue {
  airPricingDetailID: string;
  containerTypeID: string;
  containerTypeCode: string;
  containerTypeName: string;
  currencyID: string;
  currencyName: string;
  price: string;
}

export type AirPricingDetailDTOsCreate = Omit<
  AirPricingDetailDTOsFormValue,
  | 'airPricingDetailID'
  | 'containerTypeCode'
  | 'containerTypeName'
  | 'currencyName'
  | 'price'
> & {
  priceAirPricingDetail: string;
};

export type AirPricingDetailDTOsUpdate = Omit<
  AirPricingDetailDTOsFormValue,
  'containerTypeCode' | 'containerTypeName' | 'currencyName'
>;

export interface AirPricingFeeFormValue {
  airPricingFeeID: string;
  feeID: string;
  feeName: string;
  feeNo: string;
  currencyID: string;
  currencyName: string;
  unitID: string;
  internationalCode: string;
  price: string;
}
// ...
export interface UpdateStatus {
  id: string;
  status: string;
}